Compile ui files into executable
The XML files in the ui folder are now compiled into the final
executable so that they don't need to be in the correct folder
any more. If they are changed, they are automatically recompiled
on running the make command. GLib needs to be installed as it
provides the glib-compile-resources program.